added: materialized LPStakingBanner
